+++ comics/accounts/models.py
@@ -18,7 +18,8 @@ def make_secret_key():
 
 
 class UserProfile(models.Model):
-    user = models.OneToOneField(User, related_name='comics_profile')
+    user = models.OneToOneField(
+        User, on_delete=models.CASCADE, related_name='comics_profile')
     secret_key = models.CharField(
         max_length=32, blank=False, default=make_secret_key,
         help_text='Secret key for feed and API access')
@@ -36,8 +37,8 @@ class UserProfile(models.Model):
 
 
 class Subscription(models.Model):
-    userprofile = models.ForeignKey(UserProfile)
-    comic = models.ForeignKey(Comic)
+    userprofile = models.ForeignKey(UserProfile, on_delete=models.CASCADE)
+    comic = models.ForeignKey(Comic, on_delete=models.CASCADE)
 
     class Meta:
         db_table = 'comics_user_profile_comics'