+++ comics/core/command_utils.py
@@ -1,29 +1,17 @@
 import logging
-from optparse import make_option
 
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
 DATE_TIME_FORMAT = '%Y-%m-%d %H:%M:%S'
-FILE_LOG_FORMAT = '%(asctime)s [%(process)d] %(name)-12s %(levelname)-8s ' \
-    + '%(message)s'
+FILE_LOG_FORMAT = (
+    '%(asctime)s [%(process)d] %(name)-12s %(levelname)-8s %(message)s')
 CONSOLE_LOG_FORMAT = '%(levelname)-8s %(message)s'
 
 
 class ComicsBaseCommand(BaseCommand):
-    if not [option for option in BaseCommand.option_list
-            if option.dest == 'verbosity']:
-        option_list = BaseCommand.option_list + (
-            make_option(
-                '-v', '--verbosity', action='store', dest='verbosity',
-                default='1', type='choice', choices=['0', '1', '2'],
-                help=(
-                    'Verbosity level; 0=minimal output, 1=normal output, '
-                    '2=all output')),
-        )
-
     def handle(self, *args, **options):
-        self._setup_logging(int(options.get('verbosity', 1)))
+        self._setup_logging(options['verbosity'])
 
     def _setup_logging(self, verbosity_level):
         logging.root.setLevel(logging.NOTSET)
@@ -42,10 +30,10 @@ class ComicsBaseCommand(BaseCommand):
         console = logging.StreamHandler()
         if verbosity_level == 0:
             console.setLevel(logging.ERROR)
-        elif verbosity_level == 2:
-            console.setLevel(logging.DEBUG)
-        else:
+        elif verbosity_level == 1:
             console.setLevel(logging.INFO)
+        else:
+            console.setLevel(logging.DEBUG)
         formatter = logging.Formatter(CONSOLE_LOG_FORMAT)
         console.setFormatter(formatter)
-        logging.getLogger('').addHandler(console)
+        logging.root.addHandler(console)