+++ comics/core/management/commands/comics_addcomics.py
@@ -1,16 +1,15 @@
 from comics.core.comic_data import ComicDataLoader
-from comics.core.command_utils import ComicsBaseCommand, make_option
+from comics.core.command_utils import ComicsBaseCommand
 
 
 class Command(ComicsBaseCommand):
-    option_list = ComicsBaseCommand.option_list + (
-        make_option(
+    def add_arguments(self, parser):
+        parser.add_argument(
             '-c', '--comic',
             action='append', dest='comic_slugs', metavar='COMIC',
             help=(
                 'Comic to add to site, repeat for multiple. ' +
-                'Use "-c all" to add all.')),
-    )
+                'Use "-c all" to add all.'))
 
     def handle(self, *args, **options):
         super(Command, self).handle(*args, **options)